Create Report Component
